Create A Form To Add Books and Append Them To Screen
Current commit deliveres a basic program where the user can add books to the library (screen) using add book button which is associated with a form with a display: none initially, once the book is created, it is appended to the screen and a delete book button associated with this book is created. Once the delete button is pressed; the book will be deleted from the books (library) array and removed from screen.