Add Read Status Toggle
A Toggle (two buttons) is created to reflect the status of the book added, wheather it is read or unread. The status is added to the constructor and the array (library), and two buttons are created. The right button will light up once the book is added, if the user wants to change the book's status, it will also be changed in the array. Currently, the two buttons will be still there if the book is deleted.